\name{mlemur-package}
\alias{mlemur-package}
\docType{package}
\title{mlemur: MLE Mutation Rate calculator}
\description{
  A tool for calculating maximum likelihood-based mutation rates,
  confidence intervals, and P values. Offers graphical user interface built with shiny.
}
\details{
  based on rSalvador by Qi Zheng; GPL 2.0
}
\author{
Krystian Lazowski, klazowski@ibb.waw.pl
Qi Zheng

Maintainer: Krystian Lazowski <klazowski@ibb.waw.pl>
}
