% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rluria}
\alias{rluria}
\title{Simulate fluctuation test}
\usage{
rluria(
  n = 10,
  rate = 1e-08,
  N0 = 1,
  Nt = 1e+09,
  mut_fit = 1,
  death_prob = 0,
  lag = 0,
  e = 1,
  cv = 0,
  trim = 0,
  ret = "list"
)
}
\arguments{
\item{n}{number of parallel cultures in the experiment; a positive integer.}

\item{rate}{average mutation rate; a positive number smaller than 1.}

\item{N0}{size of inoculum; a positive integer.}

\item{Nt}{final number of cells in culture; a positive integer.}

\item{mut_fit}{relative fitness of the mutant cells vs. wild-type cells; a positive number. Should not be used together
with lag.}

\item{death_prob}{death probability (same for mutants and wild-type cells); a non-negative number smaller
than 0.5; relative (d)eath rate and death (p)robability are connected by the relation d = p/(1-p); p = d/(1+d).}

\item{lag}{mean phenotypic lag of mutant cells in generations; a non-negative number. Should not be used together
with mut_fit.}

\item{e}{plating efficiency; a positive number not bigger than 1.}

\item{cv}{coefficient of variation of the number of cells in culture; a non-negative number usually smaller than 1.}

\item{trim}{a non-negative integer; if specified, any mutant cell count bigger than this number will be replaced
with this number.}

\item{ret}{return parameter; either "list" (default) or "vector".}
}
\value{
either a list of length two, each containing a vector of length n: "$mc" containing colony counts in parallel
cultures while "$nt" contains the final number of cells in each culture; or a vector identical to "$mc".
}
\description{
This function simulates a fluctuation assay consisting of n cultures, starting with N0 wild-type cells
growing exponentially until reaching Nt wild-type cells. If cv is not 0, the number of wild-type cells per culture
is drawn from gamma distribution. Wild-type cell growth is taken as deterministic while mutant cell growth is taken
as stochastic according to a simple birth-and-death process. If plating efficiency is less than perfect, the success
of plating each mutant cell is simulated using binomial distribution.
}
\examples{
rluria()
rluria(n=100, rate=1e-7, lag = 2, trim=5000, ret="v")
rluria(n=50, rate=1e-9, cv=0.5)
}
\references{
Ycart B, Veziris N. Unbiased estimation of mutation rates under fluctuating final counts. PLoS One.
2014;9. doi:10.1371/journal.pone.0101434

Zheng Q. A second look at the final number of cells in a fluctuation experiment. J Theor Biol.
2016;401: 54–63. doi:10.1016/j.jtbi.2016.04.027

Mazoyer A, Drouilhet R, Despréaux S, Ycart B. Flan: An R package for inference on mutation models. R J.
2017;9: 334–351. doi:10.32614/rj-2017-029

Zheng Q. rSalvador: An R package for the fluctuation experiment. G3 Genes, Genomes, Genet.
2017;7: 3849–3856. doi:10.1534/g3.117.300120
}
